code refractor, saving and displaying user data
