+++ src/renderer/components/RouterComponents/Description/Description.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import './Description.scss';
 import { useDispatch, useSelector } from 'react-redux';
 import { displayUserData, selectCurrentUser } from '../../../../shared/slices/currentUserSlice';
-import { getUserOctokitData } from '../../../../main/git/gitCurrentUser';
 import {
   installGatsbyCLI,
   selectGatsbyInstallStatus,
@@ -24,7 +23,10 @@ const Description = () => {
       <button onClick={() => dispatch(installGatsbyCLI())}>
         Install gatsby-cli
       </button>
-      { <button onClick={() => dispatch(displayUserData(currentUser.auth.accessToken.value))}>Display user data</button>}
+      <button onClick={() => dispatch(displayUserData(currentUser.auth.accessToken.value))}>Display user data</button>
+      <p>Welcome <b>{currentUser.data.nick}</b>!</p>
+      <img src={currentUser.data.avatar} />
+      <p>Your work in <b>{currentUser.data.company}</b></p>
     </div>
   );
 };
